Place `ProxiedImage` with other image endpoints
