+++ openverse-api/catalog/api/views/site_views.py
@@ -1,8 +1,6 @@
 import logging as log
 import secrets
 import smtplib
-from urllib.error import HTTPError
-from urllib.request import urlopen
 from django.core.mail import send_mail
 from rest_framework.response import Response
 from rest_framework.reverse import reverse
@@ -14,20 +12,11 @@ from catalog.api.serializers.error_serializers import (
     ForbiddenErrorSerializer,
     InternalServerErrorSerializer,
 )
-from catalog.api.serializers.image_serializers import ProxiedImageSerializer
 from drf_yasg.utils import swagger_auto_schema
-from catalog.api.models import (
-    Image,
-    ThrottledApplication,
-    OAuth2Verification,
-)
-from catalog.api.utils.throttle import (
-    TenPerDay, OnePerSecond, OneThousandPerMinute
-)
+from catalog.api.models import ThrottledApplication, OAuth2Verification
+from catalog.api.utils.throttle import TenPerDay, OnePerSecond
 from catalog.api.utils.oauth2_helper import get_token_info
-from catalog.settings import THUMBNAIL_PROXY_URL, THUMBNAIL_WIDTH_PX
 from django.core.cache import cache
-from django.http import HttpResponse
 from drf_yasg import openapi
 from catalog.example_responses import (
     register_api_oauth2_201_example,
@@ -328,42 +317,3 @@ class CheckRates(APIView):
         }
         return Response(status=200, data=response_data)
 
-
-class ProxiedImage(APIView):
-    """
-    Return the thumb of an image.
-    """
-
-    lookup_field = 'identifier'
-    queryset = Image.objects.all()
-    throttle_classes = [OneThousandPerMinute]
-    swagger_schema = None
-
-    def get(self, request, identifier, format=None):
-        serialized = ProxiedImageSerializer(data=request.data)
-        serialized.is_valid()
-        try:
-            image = Image.objects.get(identifier=identifier)
-        except Image.DoesNotExist:
-            return Response(status=404, data='Not Found')
-
-        if serialized.data['full_size']:
-            proxy_upstream = f'{THUMBNAIL_PROXY_URL}/{image.url}'
-        else:
-            proxy_upstream = f'{THUMBNAIL_PROXY_URL}/{THUMBNAIL_WIDTH_PX}'\
-                             f',fit/{image.url}'
-        try:
-            upstream_response = urlopen(proxy_upstream)
-            status = upstream_response.status
-            content_type = upstream_response.headers.get('Content-Type')
-        except HTTPError:
-            log.info('Failed to render thumbnail: ', exc_info=True)
-            return HttpResponse(status=500)
-
-        response = HttpResponse(
-            upstream_response.read(),
-            status=status,
-            content_type=content_type
-        )
-
-        return response
